Performance: Do not draw the uncollased tree during loading
So far, the bootup phase loaded the tree once with all nodes expanded
and immediately afterwards redraw it with all nodes properly collapsed.
This was done since both coloring and collapsing the tree relied on the
`parent` property of the nodes which was set but d3 during the first
rendering of the tree.

This commit changes the logic such that we now create the parent
property ourself, without relying on d3. Thereby, we are now able to
skip the first tree layout and can immediately layout the tree we
actually want to draw.

For the dint7.xml example, this change brings the initial rendering
time down from 600-650ms to 160-180ms.