Refactoring: Cleanup interface to graph rendering
This is pretty much a refactoring commit. This commit contains the
following non-user-facing changes:
* document the internal representation of tree nodes
* move federated coloring logic out of prepareTree and only call it for
  Tableau query graphs
* So far, drawQueryTree inspected the tree and chose the node's symbol
  and CSS classes for nodes and edges by interpreting Tableau's query
  properties directly. E.g., drawQueryTree knew about federated queries
  and how to assign CSS classes to the nodes based on this.
  With this commit, drawQueryTree no longer infers symbols & CSS classes
  for nodes and edges directly. It relies on `symbol`, `nodeClass` and
  `edgeClass`. This way, the individual loaders are able to take
  decisions such as node coloring and used symbols independently of each
  other.
* collapsing is now moved into the loaders. This way, Hyper plans and
  Tableau plan can use independent logic to collapse nodes. For now,
  both just use the same duplicated logic.

The only user-facing change is that Hyper plans no longer contain
the `class`=`relation` property for tablescans, virtualtable and a few
others. So far, this property was used to make sure to render the node
with the table symbol. Since choosing the symbol is now independent from
the `class` property, we no longer need to set it. In addition, the
`class`=`relation` was displayed in the tool-tip. The class=relation is
no longer in the tooltip.
Similarly, for joins the `class` and the `join` properties are no longer
used.