More compact rendering of expressions within Hyper plans
While the previous changes made it easier to see the overall plan
structure by moving information of lower interest from the tree into
the tooltips, the related changes of the query plan loading made some
expresssion trees harder to read.

E.g., expression trees now had explicit `left` and `right` nodes in the
tree for their arguments and IUrefs were no longer labeled with the
IU name.

This commit fixes the overall algorithm such that explicit nodes are
inserted only if a we encountered an unexpected key in the JSON plan.
Since, `left` and `right` are well-known keys, this means that we don't
add them as explicit nodes. On the other hand, we now create a node for
uncommon keys such as `magic`.

In addition, this commit contains a few minor fixes such as using the
IU name as the label for iurefs.