Fix edge labels for vertical tree layouts
Edge labels were broken for all layouts except top-to-bottom because
orientation.x/y were not called. Also, the node's x0/y0 were already
updated before using them for the enter-animation of the edge labels.
Thereby, the edge labels were animated in from the parent node's end
position instead of its position at the start of the animation.

In order to make sure that this type of mistakes are harder to make in
the future, I changed when we are switching between the logical and the
display coordinate space. Before this change node.x/y were refering to
logical coordinates while x0/y0 were in the display coordinate system.
Now, they are both in the display coordinate system. Thereby, it is no
longer necessary to transform the coordinates using orienation.x()/y()
for each individual SVG element drawn.