Better collapsing for Hyper plans
So far, the keys which should be collapsed in a Hyper plan were
hardcoded. Since a few (uncommon) operators like tablefunction and
explicitscan were not taken into account, those operators were
displayed poorly.

With this change, we no longer ship a hardcoded list of keys which
should be collapsed. Instead, we now:
* for expressions: collapse no children at all. This is the same
  behavior as before.
* for operators: collapse all children except for a few selected
  children such as "left" and "right".

With this change, we now display also uncommon operators nicely.
At the same time, we will not have to update the lists of collapsed
keys when Hyper devs add new functionality.