Upgrade to D3 version 5
This commit upgrades our D3 dependency from version 3.5.16 to 5.9.2.
The largest changes were:
* the tree layout now works on a hierarchy created through
  `d3.hierarchy`. This requires some plumbing to get this hierarchy work
  with our loaders.
  I did not want to introduce a dependency to `d3.hierarchy` directly
  into the loaders, hence all changes are done in `tree-rendering.js`
* `.enter().append()` no longer adds newly added elements to the default
  selection. This required some calls to `.merge` to retain the old
  behavior
* switching to d3-fetch: this means, older browsers won't be supported
  anymore. I think that should be fine, given https://caniuse.com/#search=fetch

D3 version 4 introduced modularization of D3. This allows for smaller
bundles. However, our build is currently not setup to take advantage of
this modularization and hence the bundle size actually increases.
* min.js before: 297 kB
* min.js after:  401 kB (gzipped: 123 kB)

This will be addressed in follow-up commits which reduce file size again