Update README files to reflect the new project structure
With this commit, the individual subprojects get individual READMEs.
This is important such that the published packages on the npm registry
include a nice README.

Also, this commit documents a few tricks regarding pulling in
unpublished versions of the query-graphs library into the stand-alone
server and into the JupyterLab extension.

Last but not least: this commit adds a screenshot of the JupyterLab
extension in action.