Display table names for tablescans in Hyper plans
So far, table scans in Hyper query plans were only labeled as
`tablescan`. This was fine since most queries generated by Tableau
mostly referenced the `Extract` table anyway.
Now that we are seeing more and more normalized schemas in production,
it turns out that knowing which tables are joined is pretty crucial
information.

This commit changes our rendering of tablescans so that we always
display the name of the scanned table as the label of tablescan nodes.

While at it, I also added nicer rendering for the operators `csvscan`
and `binaryscan`.