Quick-and-dirty port of the `query-graphs` core to typescript
This commit ports the `query-graphs` core library to TypeScript.
The code was only minimally modified such that it compiles successfully
with the Typescript compiler.

This means that most places are still using `any`-typed variables. A
future commit will add stricter types.

But even with this rudimentary and barely typesafe approach, we already
get some benefits from TypeScript: E.g., it is no longer possible to
accidentally invoke a non-existing function due to a typo.