Simplify file structure
So far, the upload-server played a small trick: it virtually combined
the "uploads", the "media" and the "webroot" folder into one virtual
folder.
This trick doesn't work with "static file hosting" services, such as a
plain Apache server or with Github pages.
In preparation to take more advantage of Github pages, this commit
hence, removes this hack and instead reorganizes our repository, such
that all files are located in the "webroot" folder.

Furthermore, this commit simplifies path handling in both the frontend
and the backend. Among others, I was able to get rid of the "path"
depedency of our frontend