Enable more lint rules and fix them...
This commit enables another set of recommended linting rules. Most of
those rules we already comply with anyway.
We violated the "no-useless-escapes" rule which is now fixed by this
commit (https://eslint.org/docs/rules/no-useless-escape).
Futhermore, we are still violating the "no-prototype-builtins" rule.
This will be implicitely fixed as we move on to a more typesafe
usage of TypeScript.