Type safety for the standalone frontend
With this commit, we enable strict typesafety for our standalone
frontend. To do so, this commit:
1. replaces the `querystring` library by the builtin `URLSearchParams`.
   I couldn't find type definitions for the `querystring` library, as
   such it seems a good time to move on to the more modern,
   standard-compliant `URLSearchParams`
2. Replaces `d3-fetch` by the native `fetch` function. `d3-fetch`
   only provided type definitions for the top-level module but not for
   the direct include of "d3-fetch/src/text". Including the complete
   d3-fetch module was not an option, as it would have dragged in
   additional, unneded stuff and thereby increased our bundle size.
3. Rewrites the usage of the `properties` parameter to be typesafe.
   So far, we did not check that we are actually getting a value
   string-to-string map as parameter. Typescript wasn't happy with that.
   Hence, we are now actually verifying all our types.
4. Replace `spin` by `spin.js`. `spin` is rather outdated and
   unmaintained. It did not provide Typescript definitions. `spin.js`
   has a compatible interface, is actively maintained and provides type
   definitions

In addition, a few minor type changes to our code structure were
necessary to satisfy Typescript. E.g., type script usually has trouble
with type inference of variables which are used from within a callback
function passed to `forEach`.

To make sure we don't regress this type safety in the future, this
commit enables strict type checking in the Typescript compiler
for the complete standalone-server module.