Clean-up & update TPC-H-based Hyper query plans
This commit updates all our TPC-H-based query plans and replaces them by
the query plans as created by the latest HyperAPI version. In
particular, this is interesting for the `step2` query plan which
contained "optimizersteps".

To do so, the queries were added to our set of SQL queries in
`plan-dumper/queries`. Unfortunately, I was not sure about the license
of the TPC-H data, so this commit does not contain it :/ Everybody who
wants to regenerate those plans, will need to get his hands on some
TPC-H data set. Internally, we can simply copy those from the Hyper
repository, but I am not sure we are allowed to add them to this
repository here which is under the MIT license...

While at it, the query plans were renamed to have more speakings names:
* query1 -> tpch-q1
* query2 -> tpch-q2
* steps2 -> tpch-q2-steps
* simple -> tablescan