Consistently apply coordinate mapping directly before rendering
We distinguish between two coordinate systems:
* The logical coordinates produced by d3
* The physical on-screen coordinates which we use for rendering
and `ooo.x`/`ooo.y` map from logical coordinates to physical coordinates

So far, `node.x/y` werre logical coordinates, but `x0/y0` were physical
coordinates. This made it hard to e.g. the `diagonal` function and led
to the introduction of a `diagonalRaw` function.

With this commit, we now encapsulate x0/y0 into a subobject `prevPos`
and store those coordinates as logical coordinates. Thereby, we can:
* get rid of `diagonalRaw`
* rewrite `diagonalCrosslink` such that we no longer need
  `sourcecrosslink`/`targetcrosslink`

Furthermore, we now store the `prevPos` only at the very end of the
`update` method. Previously, we updated it somewhere in the middle of
the function. This had the unfortunate side effect, that crosslinks
picked up the new instead of the previous source position and therefore
the animation for crosslinks used to start at the wrong position