Strictly type the properties in tree description
So far, a `TreeNode` was allowed to have arbitrary members, most of
which were then rendered in the tooltip. This had two major drawbacks:
* TypeScript did not warn us if we accidentally had a typo and e.g.
  wanted to assign a value to `nodeClass` but accidentally typed `node_class`
* It was nearly impossible to introduce new members without potentially
  breaking the rendering of some plans, since you never knew which
  members were in use and which ones you could still use for a new purpose

With this change, all properties are now strictly typed. We still have 3
members `tag`, `text` and `class` which probably should be removed in
the long-term. But for now, this is good enough...