Add examples & basic support for "query locations"
Hyper recently gained the capability to annotate query plans with
backpointers to the corresponding SQL snippet from the query text.
For most base-operators (tablescans, `VALUES` clauses etc), it now
writes an `querylocs` array, containing a list of ranges. Each range is
identified by a start and an end offset, expressed in bytes from the
begining of the query text.

This commit:
* udpates the example plans in this repository
* changes the tree rendering, such that `querylocs` are displayed inside
  the tooltip instead of inside the (collapsed) subtree

Long-term, we would probably want to have a more integrated experience,
where query-graphs is able to display both the query plan and the query
text side-by-side and provides mappings between them. But that will be
something for another day...