Hyper: Properly render parquet scan & update tablescan names for recent changes to plan JSON format
Table scans and Parquet scans nowadays use the `debugName` instead of the `name` property
to store the table name. With this commit we now render the `debugName` correctly to label
the scans in the tree.

Furthermore, Parquet scans (and all other external format scans) have a `source` property.
This `source` property is structured, and hence rendered as nodes in the tree. However,
we had special logic to render `source` properties as non-collapsed by default. Hence,
the `source` properties were cluttering the structure of the query graph. With this change,
`source` properties are now rendered as collaped by default. The special treatment of `source`
was previously necessary to correctly render `explicitscan` operators which stored their
input tree in the `source` property. In a recent commit, I adjusted Hyper to use the `input`
property instead for `explicitscan`s.

So far, we rendered this structured property as nodes below the parquetscan, because

Last but not least, we now use the table scan icon for Parquet scans.