Postgres: Display "Relation Name" for all table scans
Postgres has a plethora of different scan types, and so far we
didn't render the Relation Name for almost all of them. With this
change, we now render the "Relation Name", if present.  Similarly,
for some scans, we want to display the "Index Name". Also, we no
longer list all scan types explicitely. This follows the general
design principle of "sane defaults" and makes us more robust should
Postgres add additional types of scans.