Allow node prop values to be selected
Allow the user to select the values under `qg-graph-node-body` by introducing a
new CSS class `qg-prop-value` and using it to set the CSS property `user-select`
to `all`.

This keeps all of the existing behaviour of query graph nodes, i.e. clicking it
anywhere to expand/collapse, but still makes it possible to copy out information
by right-clicking.

Not being able to do this was a bit of a nuisance sometimes.