Support passing parameters using URL fragments
So far parameters could only be passed using query strings.
Since query strings are sent to the server, URI length limitations
can apply. With this change we let the standalone app also parse
everything that comes after the hash in the url as if it were
part of the query string.